requirements.txt
boto3>=1.28.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pytest>=7.4.0
roboflow>=1.1.0
pyyaml>=6.0
